refactor: rename spotify-card to cassette-player
- Renamed `src/components/ui/spotify-card.tsx` to `src/components/ui/cassette-player.tsx` to align with the new skeuomorphic cassette player implementation.
- Updated the import path in `src/components/now-playing.tsx` to `@/components/ui/cassette-player`.
- Ran format, check, build, and visual screenshot tests to verify flawless operation.
